+++ packages/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @tmonitor/core
 
+## 1.1.2
+
+### Patch Changes
+
+- recordPlugin
+- Updated dependencies
+  - @web-tmonitor/utils@1.0.5
+  - @web-tmonitor/performance@1.1.3
+
 ## 1.1.1
 
 ### Patch Changes