+++ packages/record/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @web-tmonitor/record
 
+## 1.1.0
+
+### Minor Changes
+
+- recordPlugin
+
+### Patch Changes
+
+- Updated dependencies
+  - @web-tmonitor/utils@1.0.5
+
 ## 1.0.2
 
 ### Patch Changes